🚚(admin): move scoring simulator acceptance tests to administration tests
