--- admin/tests/acceptance/authenticated/certifications/scoring-simulation_test.js
@@ -1,64 +1,0 @@
-import { clickByName, visit, within } from '@1024pix/ember-testing-library';
-import { fillIn } from '@ember/test-helpers';
-import { setupMirage } from 'ember-cli-mirage/test-support';
-import { setupApplicationTest } from 'ember-qunit';
-import { module, test } from 'qunit';
-
-import { authenticateAdminMemberWithRole } from '../../../helpers/test-init';
-
-module('Acceptance | Route | routes/authenticated/certifications/scoring-simulator', function (hooks) {
-  setupApplicationTest(hooks);
-  setupMirage(hooks);
-
-  hooks.beforeEach(async function () {
-    this.server.create('user', { id: 888 });
-    await authenticateAdminMemberWithRole({ isSuperAdmin: true })(server);
-  });
-
-  test('should display a score and competence levels list when a capacity is given', async function (assert) {
-    // given
-    const screen = await visit(`/certifications/scoring-simulation`);
-
-    // when
-    await fillIn(screen.getByLabelText('Capacité'), 1);
-    await clickByName('Générer un profil');
-
-    // then
-
-    assert.ok(await screen.findByText('Score :'));
-    assert.ok(await screen.findByText('768'));
-    assert.ok(await screen.findByText('Capacité :'));
-    assert.ok(await screen.findByText('1'));
-
-    const table = screen.getByRole('table', { name: 'Niveau par compétence' });
-    const rows = await within(table).findAllByRole('row');
-
-    assert.dom(within(table).getByRole('columnheader', { name: 'Compétence' })).exists();
-    assert.dom(within(table).getByRole('columnheader', { name: 'Niveau' })).exists();
-    assert.dom(within(rows[1]).getByRole('rowheader', { name: '1.1' })).exists();
-    assert.dom(within(rows[1]).getByRole('cell', { name: '3' })).exists();
-  });
-
-  test('should display a competence and competence levels list when a score is given', async function (assert) {
-    // given
-    const screen = await visit(`/certifications/scoring-simulation`);
-
-    // when
-    await fillIn(screen.getByLabelText('Score global en Pix'), 768);
-    await clickByName('Générer un profil');
-
-    // then
-    assert.ok(await screen.findByText('Score :'));
-    assert.ok(await screen.findByText('768'));
-    assert.ok(await screen.findByText('Capacité :'));
-    assert.ok(await screen.findByText('1'));
-
-    const table = screen.getByRole('table', { name: 'Niveau par compétence' });
-    const rows = await within(table).findAllByRole('row');
-
-    assert.dom(within(table).getByRole('columnheader', { name: 'Compétence' })).exists();
-    assert.dom(within(table).getByRole('columnheader', { name: 'Niveau' })).exists();
-    assert.dom(within(rows[1]).getByRole('rowheader', { name: '1.1' })).exists();
-    assert.dom(within(rows[1]).getByRole('cell', { name: '3' })).exists();
-  });
-});
